+++ test/samples/operators.rb
@@ -1,1 +1,18 @@
 # Samples of different operators
+
+# range operator (..)
+0..4
+-1..-3
+'a'..'z'
+0.0..4.0
+foo..bar
+0..4.0
+0.0..4
+
+# exclusive range operator (...)
+0...4
+-1...-3
+'a'...'z'
+0.0...4.0
+foo...bar
+0...4.0